+++ test/FluentValidation.Extensions.Br.Test/Person.cs
@@ -8,5 +8,7 @@ namespace FluentValidation.Extensions.Br.Test
     {
         public string CPF { get; set; }
         public string CNPJ { get; set; }
+        public string UF { get; set; }
+        public string PISPASEP { get; set; }
     }
 }