+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [commit-conventions](https://www.conventionalcommits.org/en/v1.0.0/#specification) for commit guidelines.
 
+
+## 1.0.3 (2022-09-10)
+
+
+### 🎉 Features
+
+* **project:** including validation for uf ([aef6ef2](https://github.com/iurifarenzena/FluentValidation.Extensions.Br/commit/aef6ef2050dc4a307379ee181df3fafefb43fd30))
+* **project:** including validation for pis/pasep ([f7c81ff](https://github.com/iurifarenzena/FluentValidation.Extensions.Br/commit/f7c81ffc01ac1ed9163123b630bf862bf383b76b))
+
+
 ### 1.0.2 (2021-04-21)